Move RBF headers to include/rbf
