+++ include/rbf/gaussian.h
