+++ src/lib/radial_basis_functions/gaussian.cpp
@@ -1,1 +1,1 @@
-#include "gaussian.h"
+#include <rbf/gaussian.h>
