+++ src/lib/radial_basis_functions/interpolate.cpp
@@ -1,1 +1,1 @@
-#include "interpolate.h"
+#include <rbf/interpolate.h>
