+++ src/lib/radial_basis_functions/tests/rbf_test.cpp
@@ -1,5 +1,5 @@
-#include "../gaussian.h"
-#include "../interpolate.h"
+#include <rbf/gaussian.h>
+#include <rbf/interpolate.h>
 #include <gtest/gtest.h>
 
 #include <math_utils/math_tools.h>
